reorganize mdbt package structure
